requirements.txt
requests==2.31.0
pytest==6.2.4
pyyaml==6.0.1
colorama==0.4.4